+++ backend/app/models/inquiry.py
@@ -1,7 +1,6 @@
 import uuid
 
-from pydantic import EmailStr
-from sqlmodel import Field, Relationship, SQLModel
+from sqlmodel import Field, SQLModel
 
 # Shared properties
 class InquiryBase(SQLModel):
